Contig#getNumberOfReads() now returns a long
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1436 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ exe/org/jcvi/assembly/ace/AbacusErrorFinder.java
@@ -124,7 +124,7 @@ public class AbacusErrorFinder {
     }
     private <P extends AssembledRead, C extends Contig<P>> List<Range> getUngappedCandidateRanges(C contig) {
         
-        List<Range> gapRangesPerRead = new ArrayList<Range>(contig.getNumberOfReads());
+        List<Range> gapRangesPerRead = new ArrayList<Range>((int)contig.getNumberOfReads());
         StreamingIterator<P> readIterator = null;
         try{
         	readIterator = contig.getReadIterator();
